requirements.txt
# GlassDesk Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
authlib==1.6.0  # Enhanced OAuth implementation

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI & Processing
openai>=1.6.1
langchain==0.0.350
langchain-openai==0.0.2
chromadb==1.0.15  # Vector database for embeddings

# Database (SQLite for development, PostgreSQL for production)
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10  # PostgreSQL adapter for production
redis==5.0.1

# Background Processing
celery==5.3.4

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring and system utilities
psutil==5.9.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
black==23.11.0
flake8==6.1.0 